Make FastHeap address arithmetic 64-bit-safe (pointer hygiene, stage 3b)
pvDecommitMem rounded the block's start pointer up to a page boundary by
casting it to uint32, masking, and casting back to void*. On x64 that
round-trip discards the high 32 bits of the address, handing VirtualFree
a bogus pointer. Widen the cast to uintptr, and widen the page-align mask
to uintptr as well so it clears the low bits without zeroing the top half
of a 64-bit address. The byte-count math beside it stays uint32 (it is a
length, not an address). The page-offset difference keeps its uint32
result via an explicit narrow.

Also widen the three (int)-of-pointer alignment asserts in FastHeap.hpp
to uintptr, so the modulus checks stay meaningful (and warning-free) once
pointers are 64-bit.

Behaviour-identical on Win32 where uintptr == uint32: built green (exit 0,
no new warnings; the 18 C4291s citing FastHeap are the pre-existing
placement-new-without-delete baseline at the call sites), smoke-launched
OK. Second slice of the uint32->uintptr pointer-hygiene sweep after
VirtualMem (ba41823). The implicit pointer-difference narrowings in the
CDArray template (uCommit = (T*)pv - atArray) are left for the CDArray
slice; they are latent C4244 warnings, not truncation bugs.

+++ jp2_pc/Source/Lib/Sys/FastHeap.hpp
@@ -597,7 +597,7 @@ public:
 		// do not commit the memory in the CDArray - it will get commited as it is allocated
 		: CDArray<char>(RoundUp(u_heap_size, uHEAP_ALIGNMENT), false)
 	{
-		Assert((int)atArray % uHEAP_ALIGNMENT == 0);
+		Assert((uintptr)atArray % uHEAP_ALIGNMENT == 0);
 
 		MEMLOG_ADD_COUNTER(emlCFastHeap, u_heap_size );
 		MEMLOG_SUB_COUNTER(emlCDArray, u_heap_size );
@@ -696,7 +696,7 @@ public:
 			p_reset_position = pvGetBase();
 
 		Assert(bWithin((char*)p_reset_position, atArray, atArray + uLen));
-		Assert((int)p_reset_position % uHEAP_ALIGNMENT == 0);
+		Assert((uintptr)p_reset_position % uHEAP_ALIGNMENT == 0);
 
 		CDArray<char>::Reset((char*)p_reset_position - atArray,false,u4_decommit);
 	}
@@ -738,7 +738,7 @@ public:
 		// Remember the current position.
 		void* pv_return_value = reinterpret_cast<void*>(atArray + uLen);
 
-		Assert((int)pv_return_value % uHEAP_ALIGNMENT == 0);
+		Assert((uintptr)pv_return_value % uHEAP_ALIGNMENT == 0);
 
 		Grow(RoundUp(u_size_type, uHEAP_ALIGNMENT));
